added throw exceotion handler class and into each method where needed

+++ src/main/java/com/hogwarts/demo/CourseService.java
@@ -3,6 +3,7 @@ package com.hogwarts.demo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,16 +19,26 @@ public class CourseService {
     }
 
     public void updateGreeting(Courses newCourses, long id) {
+        if (!courseRepository.existsById(id)) {
+            throw new CourseNotFoundException();
+        }
         newCourses.setId(id);
         courseRepository.save(newCourses);
     }
 
     public Courses getCourseById(long id) {
         Optional<Courses> course = courseRepository.findById(id);
+        if (course.isEmpty()) {
+            throw new CourseNotFoundException();
+        }
         return course.get();
     }
 
+    @Transactional
     public void deleteCourseById(long id) {
+        if (!courseRepository.existsById(id)) {
+            throw new CourseNotFoundException();
+        }
         courseRepository.deleteById(id);
     }
